ZDBSP XNOD extended nodes support
This adds a function P_LoadXNOD - P_Load{Segs,Subsectors,Nodes} for
ZDBSP extended nodes - and calls it when nodes_zdbsp == 1. Certain data
types have their sizes increased to 32bit: subsectors can refer to >64k
segs, and nodes can refer to >64k child nodes.

Another code dump from rboom. A different implementation from
PrBoom-Plus's, although I did refer several times to the latter, for
fixes. It doesn't read data through packed structs (and the actual
unserialisation is a mess - I should make a set of wrapper functions or
something) Also there's less (i.e. no) error checking. Some notes:

- I thought I could get away with having two different blocks of loaded
  vertices but other code assumes they're all in one contiguous array
  (in the same way as we already have for linedefs, sectors etc.)

- Must be careful to unserialize node bounding boxes and children in the
  right order. I was effectively doing bbox1, child1, bbox2, child2 when
  in the file the order is bbox1, bbox2, child1, child2.

- P_LoadNodes (regular BSP) must account for 32-bit children; read the
  child number as an unsigned short, and translate 0xffff to 0xffffffff
  or otherwise bit 15 set (old NF_SUBSECTOR) to bit 31 (new).

- I forgot seg->iSegID which made all walls disappear in GL mode(!)

Tested with: NGmvmt1 map21, Sunder map11.

+++ prboom2/src/r_defs.h
@@ -280,7 +280,7 @@ typedef struct
 typedef struct subsector_s
 {
   sector_t *sector;
-  unsigned short numlines, firstline;
+  unsigned int numlines, firstline;
 } subsector_t;
 
 
@@ -291,7 +291,7 @@ typedef struct
 {
   fixed_t  x,  y, dx, dy;        // Partition line.
   fixed_t bbox[2][4];            // Bounding box for each child.
-  unsigned short children[2];    // If NF_SUBSECTOR its a subsector.
+  unsigned int children[2];      // If NF_SUBSECTOR it is a subsector.
 } node_t;
 
 //
